refactor(rust): extract collection export jobs (slice 7/9)
Move collection export task orchestration and CSV/JSON file-writing helpers into db/export.rs:

- start_collection_export_impl
- export_mock_collection_to_file
- export_real_collection_to_file
- task progress helpers
- CSV/JSON conversion helpers

Keep the Tauri command wrapper in lib.rs and re-export the impl from the crate root. Behavior unchanged.

cargo check: passed
cargo test: 60 passed
